adding hot reload functionality
